requirements.txt
aiosqlite==0.21.0
better_proxy==1.3.1
colorama==0.4.6
curl_cffi==0.7.4
eth_abi==5.2.0
eth_utils==5.1.0
loguru==0.7.3
py_solc_x==2.0.3
rlp==4.0.1
SQLAlchemy==2.0.37
web3==7.6.1
